Rename examples module to acceptance-tests
