+++ acceptance-tests/src/main/scala/dev/chronicles/acceptancetests/TableLoader.scala
@@ -1,11 +1,10 @@
-package dev.chronicles.examples
+package dev.chronicles.acceptancetests
 
 import java.time.Instant
 
 import cats.effect.IO
 import dev.chronicles.core.TableDefinition
 import dev.chronicles.core.VersionTracker.{UpdateMessage, UserId}
-import dev.chronicles.spark.SparkSupport
 import com.typesafe.scalalogging.LazyLogging
 import dev.chronicles.spark.{SparkSupport, VersionContext}
 import org.apache.spark.sql.{Dataset, SparkSession}
